refactor(cli): remove programmatic AST schema mutation in favor of init presets and doc snippets
