+++ apps/www/content/docs/reference/index.mdx
@@ -31,8 +31,6 @@ The `arkenv` package is the CLI. Import validation from `@arkenv/core` or
   <Card href="/docs/reference/check" title="check" description="Validate the resolved environment against your schema with CI exit codes." />
 
   <Card href="/docs/reference/example" title="example" description="Update .env.example from the schema." />
-
-  <Card href="/docs/reference/preset" title="preset" description="Apply, refresh, or remove hosting provider presets." />
 </Cards>
 
 ## Packages
