+++ apps/www/content/docs/reference/init.mdx
@@ -46,7 +46,6 @@ Pre-populate provider system variables. Values: `none`, `vercel`,
 `--host-preset` and `-H` as aliases.
 
 See [Hosting presets](/docs/core-concepts/hosting-presets).
-To apply or remove presets later, use [`preset`](/docs/reference/preset).
 
 ### `--agent`
 
@@ -116,7 +115,7 @@ Importing `arkenv` as a library throws and points you at `@arkenv/core`.
 <Cards>
   <Card href="/docs/getting-started/installation" title="Installation" description="CLI walkthrough and manual install." />
 
-  <Card href="/docs/reference/preset" title="preset" description="CLI reference for applying and removing provider presets." />
+  <Card href="/docs/core-concepts/hosting-presets" title="Hosting presets" description="Pre-populate provider environment variables." />
 
   <Card href="/docs/reference/example" title="example" description="Update .env.example from the schema." />
