+++ apps/www/content/docs/reference/meta.json
@@ -11,7 +11,6 @@
 		"init",
 		"check",
 		"example",
-		"preset",
 
 		"---Packages---",
 		"core",
