--- apps/www/content/docs/reference/preset.mdx
@@ -1,88 +1,0 @@
----
-title: preset
-description: Apply, refresh, or remove hosting provider preset blocks in your schema.
----
-
-`arkenv preset` writes managed comment blocks into your schema for a
-hosting provider. Use it to add platform variables, refresh them when
-the provider list changes, or strip a provider when you switch hosts.
-
-```package-install
-npx arkenv@latest preset apply vercel
-```
-
-```package-install
-npx arkenv@latest preset remove vercel
-```
-
-[Global flags](/docs/reference#global-flags) (`--yes`, `--quiet`,
-`--json`, `--agent`, `--help`) apply here too.
-
-## Usage
-
-```txt title="Terminal"
-arkenv preset <action> [provider] [options]
-```
-
-`action` is `apply` or `remove` (`rm` is an alias). `apply` injects or
-refreshes variables inside managed comment blocks. `remove` deletes that
-provider's block and updates `.env.example`.
-
-`provider` is `vercel`, `netlify`, `cloudflare`, `railway`, `render`, or
-`fly`. Omit it to pick interactively. With `--yes` and no provider, the
-CLI defaults to `vercel`.
-
-## Options
-
-These flags are specific to `preset`.
-
-### `--file <path>`
-
-Point at a schema file. This overrides the `"arkenv"` pointer in
-`package.json`.
-
-```package-install
-npx arkenv@latest preset apply vercel --file ./src/config/env.ts
-```
-
-### `--force` / `-f`
-
-Bypass the clean git working tree check.
-
-```package-install
-npx arkenv@latest preset apply vercel --force
-```
-
-### `--yes` / `-y`
-
-Skip confirmation prompts and accept defaults.
-
-### `--agent`
-
-Non-interactive JSON mode for agents. Same macro as the
-[global `--agent` flag](/docs/reference#agent): `--yes --quiet --json`.
-
-## Behavior
-
-The CLI detects the framework public prefix
-(`NEXT_PUBLIC_`, `NUXT_PUBLIC_`, `VITE_`), and the validator dialect
-(ArkType, Zod, Valibot). Then it:
-
-1. Reads the `"arkenv"` object in `package.json` to find schema files.
-2. Refuses a dirty git tree unless you pass `--force`.
-3. Rejects collisions with fields you defined outside managed blocks.
-4. Inserts, refreshes, or removes `// @arkenv-preset-start` /
-   `// @arkenv-preset-end` blocks.
-5. Syncs `.env.example` for added or removed keys and leaves shared keys
-   in place.
-
-See [Hosting presets](/docs/core-concepts/hosting-presets)
-for the field lists.
-
-## Next steps
-
-<Cards>
-  <Card href="/docs/core-concepts/hosting-presets" title="Hosting presets" description="Field-level breakdown of managed preset blocks." />
-
-  <Card href="/docs/reference/init" title="init" description="Scaffold a project and apply a preset in one command." />
-</Cards>
