+++ packages/arkenv/src/cli/cli.test.ts
@@ -307,87 +307,6 @@ describe("CLI parser", () => {
 			const invalid = new CLI(["node", "arkenv", "init", "-H", "vercle"]);
 			expect(invalid.validationError).toBe("Invalid host preset: vercle");
 		});
-
-		describe("preset command", () => {
-			it("should parse valid preset apply vercel command", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "apply", "vercel"]);
-				expect(cli.command).toBe("preset");
-				expect(cli.presetInput.action).toBe("apply");
-				expect(cli.presetInput.provider).toBe("vercel");
-				expect(cli.validationError).toBeUndefined();
-			});
-
-			it("should parse valid preset remove netlify command", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "remove", "netlify"]);
-				expect(cli.command).toBe("preset");
-				expect(cli.presetInput.action).toBe("remove");
-				expect(cli.presetInput.provider).toBe("netlify");
-				expect(cli.validationError).toBeUndefined();
-			});
-
-			it("should parse valid preset rm alias", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "rm", "cloudflare"]);
-				expect(cli.command).toBe("preset");
-				expect(cli.presetInput.action).toBe("remove");
-				expect(cli.presetInput.provider).toBe("cloudflare");
-				expect(cli.validationError).toBeUndefined();
-			});
-
-			it("should parse preset apply with omitted provider", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "apply"]);
-				expect(cli.command).toBe("preset");
-				expect(cli.presetInput.action).toBe("apply");
-				expect(cli.presetInput.provider).toBeUndefined();
-				expect(cli.validationError).toBeUndefined();
-			});
-
-			it("should parse flags in presetInput", () => {
-				const cli = new CLI([
-					"node",
-					"arkenv",
-					"preset",
-					"apply",
-					"vercel",
-					"--force",
-					"--yes",
-					"--file",
-					"./custom-env.ts",
-				]);
-				expect(cli.command).toBe("preset");
-				expect(cli.presetInput.action).toBe("apply");
-				expect(cli.presetInput.provider).toBe("vercel");
-				expect(cli.presetInput.isForce).toBe(true);
-				expect(cli.presetInput.isYes).toBe(true);
-				expect(cli.presetInput.file).toBe("./custom-env.ts");
-			});
-
-			it("should reject bare preset without a subcommand", () => {
-				const cli = new CLI(["node", "arkenv", "preset"]);
-				expect(cli.validationError).toBe("Missing subcommand");
-			});
-
-			it("should reject unknown preset action", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "invalid"]);
-				expect(cli.validationError).toBe("Unknown preset action: invalid");
-			});
-
-			it("should reject invalid provider in preset apply", () => {
-				const cli = new CLI(["node", "arkenv", "preset", "apply", "vercle"]);
-				expect(cli.validationError).toBe("Invalid host preset: vercle");
-			});
-
-			it("should reject extra positional arguments", () => {
-				const cli = new CLI([
-					"node",
-					"arkenv",
-					"preset",
-					"apply",
-					"vercel",
-					"extra",
-				]);
-				expect(cli.validationError).toBe("Unknown argument: extra");
-			});
-		});
 	});
 
 	describe("check command", () => {
