+++ packages/arkenv/src/cli/commands/help.test.ts
@@ -26,35 +26,19 @@ describe("HelpUseCase", () => {
 		);
 		expect(initCommandLog).toBeDefined();
 		expect(initCommandLog).toBe(
-			"  arkenv init [project-name]         Set up ArkEnv in your project",
+			"  arkenv init [project-name]    Set up ArkEnv in your project",
 		);
 
 		const checkCommandLog = logs.find((l) => l.includes("arkenv check"));
 		expect(checkCommandLog).toBeDefined();
 		expect(checkCommandLog).toBe(
-			"  arkenv check                       Validate the environment against the schema",
+			"  arkenv check                  Validate the environment against the schema",
 		);
 
 		const exampleCommandLog = logs.find((l) => l.includes("arkenv example"));
 		expect(exampleCommandLog).toBeDefined();
 		expect(exampleCommandLog).toBe(
-			"  arkenv example                     Update .env.example from the schema",
-		);
-
-		const presetApplyCommandLog = logs.find((l) =>
-			l.includes("arkenv preset apply [provider]"),
-		);
-		expect(presetApplyCommandLog).toBeDefined();
-		expect(presetApplyCommandLog).toBe(
-			"  arkenv preset apply [provider]     Apply or refresh hosting provider preset (vercel, netlify, cloudflare, railway, render, fly)",
-		);
-
-		const presetRemoveCommandLog = logs.find((l) =>
-			l.includes("arkenv preset remove [provider]"),
-		);
-		expect(presetRemoveCommandLog).toBeDefined();
-		expect(presetRemoveCommandLog).toBe(
-			"  arkenv preset remove [provider]    Remove hosting provider preset from schema",
+			"  arkenv example                Update .env.example from the schema",
 		);
 
 		const globalHeaderIndex = logs.findIndex((l) =>
