+++ packages/arkenv/src/cli/commands/help.ts
@@ -46,15 +46,6 @@ export class HelpUseCase {
 				left: "arkenv example",
 				right: "Update .env.example from the schema",
 			},
-			{
-				left: "arkenv preset apply [provider]",
-				right:
-					"Apply or refresh hosting provider preset (vercel, netlify, cloudflare, railway, render, fly)",
-			},
-			{
-				left: "arkenv preset remove [provider]",
-				right: "Remove hosting provider preset from schema",
-			},
 		];
 
 		const globalOptions: HelpItem[] = [
@@ -123,17 +114,6 @@ export class HelpUseCase {
 			},
 		];
 
-		const presetOptions: HelpItem[] = [
-			{
-				left: "--file <path>",
-				right: "Path to schema file (overrides package.json arkenv pointer)",
-			},
-			{
-				left: "--force, -f",
-				right: "Bypass dirty git working tree check",
-			},
-		];
-
 		this.logger.log(`ArkEnv CLI v${version}`);
 		this.logger.log(`\n${pc.bold("Usage:")}`);
 		for (const line of formatColumns(commands)) {
@@ -153,10 +133,6 @@ export class HelpUseCase {
 		}
 		this.logger.log(`\n${pc.bold("example options:")}`);
 		for (const line of formatColumns(exampleOptions)) {
-			this.logger.log(line);
-		}
-		this.logger.log(`\n${pc.bold("preset options:")}`);
-		for (const line of formatColumns(presetOptions)) {
 			this.logger.log(line);
 		}
 	}
