+++ packages/arkenv/src/cli/commands/index.ts
@@ -2,4 +2,3 @@ export * from "./check";
 export * from "./example";
 export * from "./help";
 export * from "./init";
-export * from "./preset";
