--- packages/arkenv/src/features/config-mutation/index.ts
@@ -1,1 +1,0 @@
-export * from "./config-mutation";
